fix(65): close branch-review tech debt (precision, harness, coverage)
Multi-agent branch review surfaced findings beyond the phase gaps; this closes them.

- time: float -> double so long deterministic runs don't lose seconds-resolution to
  float accumulation (frame/dt stay float, exact per-frame). Boundary promotions made explicit.
- client.wait_frames(): use result() so a RUN-only rejection (manual/paused -> bad_params)
  raises instead of silently returning an {ok:false} envelope.
- net: bound recv-buffer growth inside the poll loop so a loopback flood can't balloon the
  buffer (and realloc-trap) before the post-dispatch cap check.
- transport: a read timeout now invalidates the transport and normalizes to TimeoutError
  (socket.timeout is not a TimeoutError subclass on py3.8; a mid-read timeout also leaves the
  buffered reader undefined). client: cap the pending-reply stash and detect a server that
  frames batches as individual envelopes.
- nt_app.h: record the uint32 frame-wrap invariant (devapi deadlines use modular compare, < 2^31).
- tests: 5 L1-mutator assert death-tests (NT_ASSERT_FULL) + an end-to-end test that drives the
  real nt_app_run(MANUAL) loop draining time.step (the integration the hand-bumped unit ticks
  stubbed out). Strip stale nt_stats_stub / tick comments and plan tags.

Verified: 85/85 ctest, clang-format + clang-tidy clean, live devapi_host + time_render_demo.py
PASS, wait_frames RUN/pause probe PASS.

+++ engine/app/nt_app.h
@@ -16,10 +16,10 @@ typedef enum nt_app_mode_t {
 
 typedef struct nt_app_t {
     float dt;        /* Clamped delta time (seconds) */
-    float time;      /* Elapsed time since loop start (seconds) */
+    double time;     /* Elapsed seconds; double so long deterministic runs don't lose resolution to float accumulation */
     float max_dt;    /* Clamp threshold (seconds), default 0.1f */
     float target_dt; /* Frame rate cap (seconds), 0 = uncapped */
-    uint32_t frame;  /* Sim-advance counter (frozen on pause/manual-idle) */
+    uint32_t frame;  /* Sim-advance counter (frozen on pause/manual-idle); uint32 wrap is intentional — devapi deadlines use modular compare, bounded < 2^31 */
 
     /* Managed-loop time control (read/written via the nt_app_* mutators below). */
     bool paused;         /* RUN-mode pause: dt = 0, frame frozen */
